Backend-Projects/Beginner/Todo-List-API/requirements.txt
bleach==6.2.0
Flask==3.1.2
peewee==3.18.2
PyJWT==2.10.1
python-dotenv==1.1.1
py-bcrypt==0.4
